build(emails): move email pages into their own app
