+++ emails/templates/emails.html
@@ -19,7 +19,7 @@
                     <a href="outbox" class="list-group-item list-group-item-action d-flex align-items-center">
                             <span class="icon mr-3"><i class="fe fe-truck"></i></span>Outbox
                     </a>
-           
+
                     <a href="drafts" class="list-group-item list-group-item-action d-flex align-items-center">
                         <span class="icon mr-3"><i class="fe fe-file"></i></span>Drafts
                     </a>
